Upload RedHat Operator bundle automatically for full release tags
RedHat has recently changed the process for updating a new version of the operator. One of the last steps of this process was to upload the operator bundle,
which includes some metadata and manifests in the form of yaml files. This used to be a manual process by which we would upload a zip file at this location: https://connect.redhat.com/project/3928601/operator-metadata.
However, there is now an Operator Bundle project https://connect.redhat.com/project/5791371/images, where you can upload a docker container image containing the same files.