Upload RedHat Operator bundle automatically for full release tags (#45)
## Why

RedHat has recently changed the process for updating a new version of the operator. One of the last steps of this process was to upload the operator bundle, which includes some metadata and manifests in the form of yaml files. This used to be a manual process by which we would upload a zip file at this location: https://connect.redhat.com/project/3928601/operator-metadata. However, there is now an Operator Bundle project https://connect.redhat.com/project/5791371/images, where you can upload a docker container image containing the same files. Therefore, we should perform this last step automatically instead of having to remember to do this manually.

## What

Update the Jenkins pipeline to download all the assets for a specific GitHub Release (so as to not assume that those assets are available on hand, for when we do move this pipeline to Concourse and each pipeline job being run in a separate container with a clean slate). Then unzip the specific `redhat-<version>.zip` file, build a docker container containing those files in that zip directory, and upload that container image to the RedHat registry so it shows up in the `Instana-Agent-Operator-Bundle` project: https://connect.redhat.com/project/5791371/images