Modify the git pre-commit hook to check if license headers were added instead of actually adding them
# Why

In the current implementation of the git pre-commit hook, it runs `mvn validate license:format` which would automatically add license headers to relevant files that do not have them. If it passes successfully (and adds missing license headers) the commit would complete after the user types in a commit message, but wouldn't include the new changes from the license header additions. This could cause the user to incorrectly assume that the pre-commit hook had already added the necessary license headers and go on their merry way, though technically the user's PR build will fail since license headers are checked when `mvn verify` is called so they will be notified anyway, but only after waiting for their PR build to fail.

# What

* Modify the pre-commit hook to _check_ (via `mvn validate license:check`) that all appropriate license headers have been added to new files instead of actually adding them (via `mvn validate license:format`). If the check fails, the user will be directed to run `mvn validate license:format` manually to add the missing license headers
* Execute `mvn git-build-hook:install` during the `validate` phase to ensure the latest pre-commit hook is updated in `.git/hooks`
* Bind the `license:check` goal to the `validate` phase so we just need to run `mvn validate` instead of `mvn validate license:check`