fix: Adjust kubebuilder to fix cluster role
The ClusterRole produced by the kubebuilder comments when running `make
manifests` was out of sync with the definition in the
CustomResourceDefinition. Whenever the operator was started locally via
`make run`, the permissions from the local kubeconfig were inherited and
worked fine, but starting from an image inside the cluster which relied
on the serviceaccount mapped to the maanger-role ClusterRole failed with
forbidden.
This can be reproduced by running: `kubectl get
--as=system:serviceaccount:instana-agent:instana-agent-operator
instanaagent.instana.io/instana-agent -n instana-agent`