fix: If pull secret is generated, support external secret
While reading the code I noticed that the image pull secret for agent
images starting with containers.instana.io will be automatically
generated. This only worked for secrets which were encoded directly in
the CR, but it did not support the keysSecret reference yet.

This commit will ensure that the keysSecret is used first and falls back
to the deprecated direct defintion in the custom resource.

TODO: Need to re-enable the test case after fixing it, manual test
showed that secrets are picked up from the keysSecret