Fix: Bind only enabled ports on agent pods (#211)
* refactor: Remove obsolete agent-socket port 42666

* fix: Only attach enabled ports on agent pod

If opentelemetry.grpc.enabled = false or opentelemetry.http.enabled =
false, ports 55680, 4317 and 4318 should not be attached to the
Daemonset.

* ci: Stabilize e2e test, retry to cat backend files

* fixup: Exit retry loop with break

* ci: Fix e2e to check for pod name again if exec failed

If the old agent pod is still terminating, it might be picked up
incorrectly in the later test. Therefore, every failed exec should
result in fetching the pod name again.

* test: Improve ports unit test

* test: Extend test coverage of ports builder tests

* Apply suggestions from code review

Co-authored-by: Milica-Cvrkota-IBM <109787570+Milica-Cvrkota-IBM@users.noreply.github.com>

* test: Removed double check in e2e test

* fix: Tolerate legacy settings OLTP.Enabled.Enabled

If no legacy setting is defined, OLTP should be enabled by default.
If HTTP or GRPC is explicitly defined, use this value.
But, if OLTP is explicitly disabled by the legacy config, keep that
behavior and do not impliticly enable ports

* test: Refactor/simplify inline_types and unit tests

---------

Co-authored-by: Milica-Cvrkota-IBM <109787570+Milica-Cvrkota-IBM@users.noreply.github.com>
Signed-off-by: Rory Welch <Rory.Welch@ibm.com>