chore: Add support for local podman build via buildkit
If running on podman, the Dockerfile syntax is not accepted in all cases
and the build might behave differently compared to Docker. Buildkit is
the foundation of the docker build logic and can be used in isolation,
also as part of a running container. Therefore, the Makefile now detects
podman on docker on demand and installs the buildctl command line if
necessary. The build is always executed via buildctl now, which should
provide a consistent build behavior.