Add ClusterRole/ClusterRoleBinding to service-agent (#243)
In addition we needed to add some missing RBAC permissions that
are required for the instana-agent and were in the v1 version of the
the helm charts to the instana-agent-operator account since it cannot
grant RBAC permissions (to instana-agent) that itself doesn't have.

Also refactored the list of "reading verbs" from a ClusterRole
perspective i.e: "get", "list", "watch" into a function that can be used
by all rbac clusterroles.

Signed-off-by: Antonio Gutierrez <antonio.gutierrez1@ibm.com>
Co-authored-by: Antonio Gutierrez <antonio.gutierrez1@ibm.com>
Signed-off-by: Rory Welch <Rory.Welch@ibm.com>