chore: Do not pin go patch level (#253)
When reading
https://github.com/golang/go/issues/62278#issuecomment-1693538776 I
noticed that this is only recommended for development versions which are
not released yet. This is no longer valid, by default the recommendation
is to not use the patch level. If you need to pin to a explicit version,
toolchain should be used since 1.21 instead, like `toolchain go1.24.3`. Setting just `go 1.24` will avoid failures if developers do not install the latest patch level yet, but allow the build to pick the most recent patch level.

Signed-off-by: Rory Welch <Rory.Welch@ibm.com>