Fixes issues with dependencies
Fixes issues where the golang binaries might be infested with different versions of binaries causing issues due to version mismatch.
How the makefile uses binaries is no longer opt-in, so no global binaries can be referred to from the makefile, resulting in full control
of the versions in the project.