chore(deps): Remove dependency to github.com/openshift/client-go
The client-go package uses a conflicting version of
structured-merge-diff, as
k8s.io/apimachinery/pkg/apis/meta/v1/unstructured is using a different
major version. This causes problems with `go vet` and we only used the
client-go library in the e2e test suite. The library can be replaced
with this logic.