chore: Update dependency handling in Makefile, update k8s.io dependencies and controller-manger (#319)
* ci: Update dependency handling, migrate renovate config schema

Renovate was not detecting dependencies in the Makefile before. With
these adjustments, renovate will propose updates without relying on the
go.mod file necessarily.

Also, the renovate config used an outdated schema, which I updated now.

* chore(deps): Update k8s.io dependencies and controller-runtime

* chore(deps): Remove dependency to github.com/openshift/client-go

The client-go package uses a conflicting version of
structured-merge-diff, as
k8s.io/apimachinery/pkg/apis/meta/v1/unstructured is using a different
major version. This causes problems with `go vet` and we only used the
client-go library in the e2e test suite. The library can be replaced
with this logic.