chore: Migrate to go.kubebuilder.io/v4
The operator-sdk cannot be updated to latest, if we are still using
go.kubebuilder.io/v3 in our project.

I applied the manual migration steps as outlined here: https://book.kubebuilder.io/migration/manually_migration_guide_gov3_to_gov4#update-kustomize-manifests-with-the-changes-made-so-far

With these changes, the relatedImages section is rendered correctly
again in the olm bundle.