fix: use valid json for alm-examples, add other related images
The new operator-manifest-tools will search for all image references
automatically and pins them. It will not detect if tag and image name
are split, therefore adding an annotation to refer to k8s_sensor and
agent images. The relatedImages section seem no longer to be manually
maintained in the yaml, therefore removing it. Also, removing the agent
image annotation, as the operator-manifest-tools tool will detect the
deployed image from the deployment automatically and will introduce
duplicated images in the relatedImages section otherwise