fix: add instanaagent and instanaagents as shortNames for agent CRs
See https://github.com/instana/instana-agent-operator/issues/227, if the
CRD kind is requested, it is not returned as the kubebuilder annotations
do not define them as known isngular or plural. As this is used for
generating the CRDs and we want to keep changes minimal, I added a
workaround to mention the kinds as shortnames. This will support both
approaches without changing the file names/kind of the generated CRD.