flake: use nixpkgs branch 25.05 since that contains go 1.25.3
Since this is the version of the current toolchain directive in the
go.mod file, the compilation process will exit with a non-sucess exit
code when using a go root different from that one.

It seems this only affects environments using the flake.

In addition we add the flake.lock into the list of excluded files from
the detect-secrets scans, since all that has is the narhas and the rev
(commit) hash for that specific branch at that particular time.

Signed-off-by: Antonio Gutierrez <antonio.gutierrez1@ibm.com>